refactor: 방향 전략 생성 분기문 삭제
