+++ src/main/java/chess/domain/ChessRunner.java
@@ -19,7 +19,7 @@ public class ChessRunner {
         Position targetPosition = Position.of(target);
         Piece selectedPiece = this.board.getPiece(sourcePosition);
 
-        if ((currentTeam.isSameTeamWith(selectedPiece.getTeam()))) {
+        if (!(currentTeam.isSameTeamWith(selectedPiece.getTeam()))) {
             throw new IllegalArgumentException("현재 차례가 아닙니다.");
         }
