+++ src/main/java/chess/domain/board/BoardInitializer.java
@@ -1,6 +1,6 @@
 package chess.domain.board;
 
-import chess.domain.initialize.*;
+import chess.domain.strategy.initialize.*;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
