+++ src/main/java/chess/domain/strategy/initialize/KingInitializer.java
@@ -1,4 +1,4 @@
-package chess.domain.initialize;
+package chess.domain.strategy.initialize;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
@@ -9,12 +9,12 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-public final class QueenInitializer implements InitializeStrategy {
+public final class KingInitializer implements InitializeStrategy {
     @Override
     public Map<Position, Piece> initialize() {
         Map<Position, Piece> pieces = new HashMap<>();
-        pieces.put(Position.of("d8"), new Piece(PieceType.QUEEN, Team.BLACK));
-        pieces.put(Position.of("d1"), new Piece(PieceType.QUEEN, Team.WHITE));
+        pieces.put(Position.of("e8"), new Piece(PieceType.KING, Team.BLACK));
+        pieces.put(Position.of("e1"), new Piece(PieceType.KING, Team.WHITE));
 
         return Collections.unmodifiableMap(pieces);
     }
