+++ src/main/java/chess/domain/strategy/initialize/PawnInitializer.java
@@ -1,4 +1,4 @@
-package chess.domain.initialize;
+package chess.domain.strategy.initialize;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
